autonomous_dev_agent/requirements.txt
langchain>=0.3.24
langchain-core>=0.3.55
langchain-experimental>=0.3.4
langchain-community>=0.3.22
langchain-openai>=0.3.14
langchain-anthropic>=0.3.12
langgraph>=0.3.32
openai>=1.75.0
anthropic>=0.50.0
pydantic>=2.11.3
pydantic-settings>=2.9.1
python-dotenv>=1.1.0
requests>=2.32.3
pytest>=7.4.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
